+++ MiniloGUI/MiniloGUI 2026/Project_Source_Files/D_TREE_Exposed_Info_P.h
@@ -14,12 +14,12 @@ class Tree_Exposed_Info_P :
 //==============================================================================
 public: Tree_Exposed_Info_P();
 
-public: const Ctrl_Type ctrl_type_for(const uint8 param_index) const;
-public: const Knob_Display_Type knob_display_type_for(const uint8 param_index) const;
-public: const uint8 cc_num_for(const uint8 param_index) const;
-public: const uint8 bit_count_for(const uint8 param_index) const;
-public: const uint8 byte_index_for_param_bit(const uint8 param_index, const uint8 bit_index) const;
-public: const uint8 bit_index_for_param_bit(const uint8 param_index, const uint8 bit_index) const;
+public: const Ctrl_Type ctrl_type_for(const int param_index) const;
+public: const Knob_Display_Type knob_display_type_for(const int param_index) const;
+public: const int cc_num_for(const int param_index) const;
+public: const int bit_count_for(const int param_index) const;
+public: const int byte_index_for_param_bit(const int param_index, const int bit_index) const;
+public: const int bit_index_for_param_bit(const int param_index, const int bit_index) const;
 
 //==============================================================================
 private: JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(Tree_Exposed_Info_P)
