+++ MophoGUI/MophoGUI 2026/Project_Source_Files/D_TREE_Exposed_Info_P.h
@@ -17,16 +17,16 @@ class Tree_Exposed_Info_P :
 public: Tree_Exposed_Info_P();
 
 public: const Identifier id_for(Track track, Step step) const;
-public: const Ctrl_Type ctrl_type_for(const uint8 param_index) const;
-public: const Knob_Display_Type knob_display_type_for(const uint8 param_index) const;
-public: const uint8 nrpn_for(const uint8 param_index) const;
-public: const uint8 param_index_for(const uint8 nrpn) const;
-public: Point<int> red_toggle_center_for(uint8 param_index) const;
-public: const uint16 packed_bits_byte_index_for(const uint8 param_index) const;
-public: const uint8 packed_bits_mask_for(const uint8 param_index) const;
-public: const uint16 byte_index_for(const uint8 param_index) const;
-public: const Track seq_track_for(const uint8 param_index) const;
-public: const Step seq_track_step_for(const uint8 param_index) const;
+public: const Ctrl_Type ctrl_type_for(const int param_index) const;
+public: const Knob_Display_Type knob_display_type_for(const int param_index) const;
+public: const int nrpn_for(const int param_index) const;
+public: const int param_index_for(const int nrpn) const;
+public: Point<int> red_toggle_center_for(int param_index) const;
+public: const int packed_bits_byte_index_for(const int param_index) const;
+public: const int packed_bits_mask_for(const int param_index) const;
+public: const int byte_index_for(const int param_index) const;
+public: const Track seq_track_for(const int param_index) const;
+public: const Step seq_track_step_for(const int param_index) const;
 
 //==============================================================================
 private: JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(Tree_Exposed_Info_P)
