+++ MophoGUI/MophoGUI 2026/Project_Source_Files/c_EXP_p.h
@@ -5,23 +5,23 @@
 namespace EXP
 {
 
-	constexpr uint8 choice_count_bend_range{ 13 };
-	constexpr uint8 choice_count_clock_tempo{ 221 };
-	constexpr uint8 choice_count_lfo_freq{ 167 };
-	constexpr uint8 choice_count_lpf_freq{ 165 };
-	constexpr uint8 choice_count_mod_dest{ 47 };
-	constexpr uint8 choice_count_mod_src{ 23 };
-	constexpr uint8 choice_count_osc_pitch{ 121 };
-	constexpr uint8 choice_count_osc_fine{ 101 };
-	constexpr uint8 choice_count_osc_shape{ 104 };
-	constexpr uint8 choice_count_voice_name_char{ 128 };
-	constexpr uint8 clock_tempo_offset{ 30 };
-	constexpr uint8 exp_param_count{ 189 };
-	constexpr uint8 first_lfo_pitched_freq_choice{ 90 };
-	constexpr uint8 first_lfo_synced_freq_choice{ 151 };
-	constexpr uint8 first_seq_step_param_index{ 109 };
-	constexpr uint8 first_voice_name_char_param_index{ 173 };
-	constexpr uint8 param_index_seq_track_1_step_1{ 109 };
-	constexpr uint8 param_num_seq_track_1_destination{ 101 };
+	constexpr int choice_count_bend_range{ 13 };
+	constexpr int choice_count_clock_tempo{ 221 };
+	constexpr int choice_count_lfo_freq{ 167 };
+	constexpr int choice_count_lpf_freq{ 165 };
+	constexpr int choice_count_mod_dest{ 47 };
+	constexpr int choice_count_mod_src{ 23 };
+	constexpr int choice_count_osc_pitch{ 121 };
+	constexpr int choice_count_osc_fine{ 101 };
+	constexpr int choice_count_osc_shape{ 104 };
+	constexpr int choice_count_voice_name_char{ 128 };
+	constexpr int clock_tempo_offset{ 30 };
+	constexpr int exp_param_count{ 189 };
+	constexpr int first_lfo_pitched_freq_choice{ 90 };
+	constexpr int first_lfo_synced_freq_choice{ 151 };
+	constexpr int first_seq_step_param_index{ 109 };
+	constexpr int first_voice_name_char_param_index{ 173 };
+	constexpr int param_index_seq_track_1_step_1{ 109 };
+	constexpr int param_num_seq_track_1_destination{ 101 };
 
 }
