+++ _Base_Source_Files/D_SUBTREE_Choices_B.h
@@ -6,8 +6,8 @@
 struct Subtree_Choices_B
 {
 	static ValueTree off_on(const bool curt = false);
-	static ValueTree osc_pitch(uint8 choice_count, const bool curt = false);
+	static ValueTree osc_pitch(int choice_count, const bool curt = false);
 	static ValueTree signed_7_bit_int(const bool curt = false);
-	static ValueTree unsigned_int(uint16 choice_count, const bool curt = false);
+	static ValueTree unsigned_int(int choice_count, const bool curt = false);
 };
